gdk: Export GDK_KEY constants as u32.
Key values are generally guint or guint32 in API calls,
but gir defines the GDK_KEY_* contants as gint, forcing
every use to be cast. This is probably just the generator
assuming gint for #define contants, and the signed vs.
unsigned distinction doesn't matter in C or most scripting
languages. But it does matter in rust.

Special-casing these in the the gir generator, or marking
up the gdk C header to add a U literal specifier everywhere
is the ideal fix, but in the meantime, we can hide the
bug by casting when we export the ffi version of the constants.