cairo: ImageSurface: return a Result<ImageSurface, Status> from the creation functions
When Cairo gets asked to create a surface, it will always return a
non-null cairo_surface_t*, but its cairo_surface_status() may not be
CAIRO_STATUS_SUCCESS.

For the Rust-side surface creation functions, we now return a
Result<ImageSurface, Status>.  If Ok(), then the surface is in
Status::Success.  If Err(), we return the status from Cairo's error
surface.

This means that "image_surface = from_glib_full(ptr)" needs to be
passed a surface that is *not* in error.  This function will panic if
passed a Cairo surface that is in an error state.