glib: glib_wrapper!() - Specify ffi class in Object<ffi:Name, ffi:ClassName>
This is an API break for users of the glib_wrapper!() macro.

Now in the Wrapper trait we keep two associated types:

pub trait Wrapper {
    type GlibType: 'static;
    type GlibClassType: 'static;   // new one
}

This keeps the types of the instance struct and the class struct
together, so we can access them as needed.  We will need to access
both in the code generator for GObject boilerplate.

This is for https://github.com/gtk-rs/glib/issues/233