cairo: Rewrite how refcounting and pointer conversions work for Pattern
It was unsafe before as it could "borrow" the pointer owned by a
context without ensuring that the context actually stays in scope and
had various functions that (unsafely without marked as such) change if
and how the Pattern would be dropped or not.
This caused spurious crashes in the testsuite.

Also rename pointer functions to be more consistent with how they are
called elsewhere and implement Clone and Debug for the patterns.