gtk: Acquire the default MainContext when initializing GTK
It is required for GTK to be able to own the default main context on the
thread that initialized it, among other things to run its main loop on
that.

Make this more explicit by already acquiring it here during
initialization. This also ensures that other API that checks that this
very thread owns the main context works out of the box.