gtk: Remove pre-init assertion to check for non-debug GTK builds
While previously checking for this could be used to determine if GTK was
built with assertions enabled or not, this is not true anymore with
later versions of GTK. There is no way to know if assertions were
disabled or not at runtime.

While disabling assertions easily causes memory unsafety in GTK code,
it is nothing that can be disabled accidentally. Users deploying and
running such code will have to live with the consequences.

Fixes https://github.com/gtk-rs/gtk/issues/821