glib: Don't ever get a strong reference to the main context channel GSource from a Sender
Otherwise we would unref it potentially from another thread and then
cause the callback of the source to be dropped from a different thread
than where it was created. This would be wrong as the callback closure
is not Send!

Do this by counting the number of senders separately instead of re-using
the strong count of the channel Arc.