glib: Refactor futures executor GSource to ensure that futures are never dropped from a different thread
The TaskSource now only keeps the Future that has to be polled and a
reference to a Waker, which is a child source of the TaskSource.

This Waker is passed as waker to the future's poll() function and is
potentially passed to other threads and might also outlive the
TaskSource but its only job is to wake up the TaskSource.

The TaskSource is never passed to different threads and is internal
to the futures executor.

Also remove the check/prepare functions from the GSources as they're not
needed when using child sources or triggering the GSources via setting
the ready time.