gtk: Make AccelGroup::connect() and ::connect_by_path() more usable
By renaming it to connect_accel_group() for preventing conflicts with
Object::connect() and by taking a normal Rust closure with fixed types
as arguments instead of a generic glib::Closure.

Return the generated glib::Closure as it is required for usage with a
few of the other AccelGroup and AccelLabel functions.

Fixes https://github.com/gtk-rs/gtk/issues/913