glib: Remove Borrowed variants from boxed/shared/object/string wrappers
With this shared/object are only going to occupy a single pointer
instead of two. Boxed/GString still need to be an enum and will occupy two
pointers because they need to distinguish Rust allocated and GLib
allocated memory.

Also make GString enum private, it's an implementation detail that we
should not expose.