gdk-pixbuf: Revamp argument checks in Pixbuf::new_from_mut_slice and ::put_pixel
For new_from_mut_slice, we change the condition to "data.len() >= ..."
instead of "==" so that extra data after the last row gets
accepted (GdkPixbuf allows this).  Normally rows look like

   RGB(A)-data... padding
   \------rowstride-----/

But the last row can be shorter than the rowstride, i.e. it can be
only the RGB(A) data without padding.  This commit adds a test that
the argument checks in Pixbuf::new_from_mut_slice take this into
account.

This commit also changes Pixbuf::put_pixel() to use u32 for the x/y
arguments, to make an argument check for >= 0 superfluous.