glib/translate: take advantage of Rust Option & Result types
This commit introduces OptionToGlib and TryFromGlib to help with
translations from and to Glib types which encode a value with a
sentinel to indicate a None value and/or for which some values
are invalid.

These traits allow auto-implementing Option<T>::to_glib,
Option<T>::from_glib or Result::<Option<T>, _>::from_glib from
dependent crates when appropriate despite the orphan rule.