In glib_object_wrapper!, use generic Class type instead defining a struct
This makes the `$rust_class_name` argument in `glib_wrapper!` optional.
If specified, it defines a `type` alias for the class. At least
optionally allowing it is required to avoid breaking lots of gir
generated code, but perhaps it can be removed after a change to gir.

This removes the `ObjectType::RustClassType` associated type, since it
can't be enforced/assumed by the type system that that is an instance of
this generic struct.

The `IsClassFor` trait is also removed, superseded by this generic
type.

This also changes the `IsSubclassable` trait to be implemented on the
object struct rather than the class struct, since there otherwise would
be foreign trait on foreign type errors.