Improve COPYRIGHT / LICENSE files
This follows the general approach of the Rust project.

Especially don't include bogus copyright information with non-existing
legal entities, or completely outdated copyright information but instead
explicitly state that the copyright of each contributor is retained and
the version control system can be used to get the full authorship
information.

Also rewrite the paragraph about the licenses of GNOME libraries to be
less specific and simply refer to their license information instead of
pointing to a non-existing LGPL file.