CI: add CI for Windows
Add a windows workflow, which builds the creates with MSVC.
The job builds the gtk dependencies using gvsbuild and caches them
so that they are not rebuilt on each run.
For now the job just builds with the stable toolchain and x86_64 arch,
since we are doing a full matrix with different rust versions on
Linux and doing the same for Windows seems excessive.
Many steps are currently failing (eg ABI checks), so they are disabled,
however these are also failing on Linux, so it does not seem
Windows specific.