glib: Distinguish between classes and interfaces at the type level
Without this it was possible to cast from a Class<T> to a Class<U> where
U is an interface type. This doesn't work as interface structs are to be
retrieved differently than just casting and would cause memory unsafety
issues.

And implement a bit more API for working with them.

Fixes https://github.com/gtk-rs/gtk-rs/issues/348