gio: do not take args in application.run()
Passing the args is a C-ism of the API since in C the args are
passed to main(), but in rust they are in std::env, so we can
keep the common case simple.
A run_with_args variant is added for the cases where a caller
wants to process the arg vector before running the app.